FIXED BUG 2
