+++ Frontend/src/components/CreateRoomModal.jsx
@@ -6,7 +6,7 @@ import SigninPage from "./SignInPage.jsx";
 import Overlay from "./Overlay.jsx";
 
 
-const CreateRoomModal = ({ onClose }) => {
+const CreateRoomModal = ({ onClose,toggleSignupOverlay }) => {
   const [creator, setCreator] = useState("");
   const [topic, setTopic] = useState("");
   const [participantsLimit, setParticipantsLimit] = useState(1);
@@ -85,7 +85,7 @@ const CreateRoomModal = ({ onClose }) => {
       <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50 z-50">
         <SigninPage
           onClose={onClose}
-          toggleSignupOverlay={() => console.log("Open Signup Overlay")}
+          toggleSignupOverlay={toggleSignupOverlay}
           
           
         />
