+++ Frontend/src/components/Overlay.jsx
@@ -6,10 +6,36 @@ import SignupPage from '../components/SignUpPage.jsx';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import CTA_Section from './CTA_Section.jsx';
+import CreateRoomModal from "./CreateRoomModal.jsx"
+import JoinRoomModal from "./JoinRoomModal.jsx"
+import HeroSection from './HeroSection.jsx';
+import { useNavigate,useLocation } from 'react-router-dom';
+import AOS from 'aos';
 const Overlay = () => {
   const [user, setUser] = useState(null); // User state
   const [showSignin, setShowSignin] = useState(false);
   const [showSignup, setShowSignup] = useState(false);
+  const [showCreateRoom, setShowCreateRoom] = useState(false);
+  const [showJoinRoom, setShowJoinRoom] = useState(false);
+  const navigate = useNavigate();
+  const location = useLocation();
+
+  const toggleCreateRoomModal = () => {
+    setShowCreateRoom(!showCreateRoom);
+  };
+
+  const toggleJoinRoomModal = () => {
+    setShowJoinRoom(!showJoinRoom);
+  };
+
+  const isRoomPage = location.pathname.startsWith('/rooms/');
+
+  useEffect(() => {
+    AOS.init({
+      duration: 1000,
+      once: true,
+    });
+  }, []);
   useEffect(() => {
     const unsubscribe = auth.onAuthStateChanged(setUser);
     return () => unsubscribe(); 
@@ -44,6 +70,15 @@ const Overlay = () => {
         user={user} // Pass the user state to Header
         onSignOut={handleSignOut} 
       />
+      {/* Render Modals */}
+      {showCreateRoom && <CreateRoomModal onClose={toggleCreateRoomModal} toggleSignupOverlay={toggleSignupOverlay} navigate={navigate} />}
+      {showJoinRoom && <JoinRoomModal onClose={toggleJoinRoomModal} toggleSignupOverlay={toggleSignupOverlay} navigate={navigate} />}
+
+
+      {/* Hero section */}
+      <HeroSection onCreateRoomClick={toggleCreateRoomModal} onJoinRoomClick={toggleJoinRoomModal} />
+      
+
 
       {/* Signin Overlay */}
       {showSignin && (
